Sort Array
Given an array of numbers, sort the numbers ascending (1,2,3...) and return the sorted array.

In fact, returning any negative value will have the same effect as returning -1 here. Any positive value will do the same as 1. This means we can shorten this function significantly:

[3,2,4,1].sort(function comparison(a,b) {
    return a - b;
});
Here, if a is less than b, the result of the subtraction will be negative (placing a first). If b is less than a the result will positive (placing b first).